Move to MIT licence
This repository was created to host example code to encourage use of the
SensorThingsAPI.  While some of the functions are reusable and could act
as a library (e.g. query_api, parse_datastream), we have decided that
the main purpose of this code was for users to modify it.  The
requirements of the LGPL licence added constraints on the distribution
of this modified code that may discourage doing so.  As we are happy for
modified code to be redistributed with attribution, from now on, the
code will be availble under the MIT licence.